// tauri/src/openapi.d.ts
export interface components {
    schemas: {
        BaseUser: {
            /** Format: uuid */
            readonly id: string;
            first_name: string;
            last_name: string;
            /** Format: email */
            email: string;
            team_name: string;
            avatar_url?: string | null;
            /** @default false */
            is_admin: boolean;
            /** @description Whether the user is currently active (connected via websocket) */
            is_active?: boolean;
            /** Format: uint */
            team_id?: number;
            /** Format: date-time */
            readonly created_at?: string;
            /** Format: date-time */
            readonly updated_at?: string;
        };
        PrivateUser: components["schemas"]["BaseUser"] & {
            social_metadata?: {
                [key: string]: unknown;
            } | null;
            metadata?: {
                [key: string]: unknown;
            } | null;
        };
        Error: {
            message?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}